refactor(conversion): route parity and VLM tasks by domain
