+++ scripts/inference/vlm_generation_utils.py
@@ -1,0 +1,343 @@
+# Copyright (c) 2026, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Shared utilities for the canonical VLM generation entry point."""
+
+import io
+import logging
+from typing import Optional
+
+import torch
+from PIL import Image
+
+from megatron.bridge.utils.safe_url import is_safe_public_http_url, safe_url_open
+
+
+logger = logging.getLogger(__name__)
+
+
+try:
+    from qwen_vl_utils import process_vision_info
+
+    _HAS_QWEN_VL_UTILS = True
+except ImportError:
+    _HAS_QWEN_VL_UTILS = False
+
+
+def patch_kimi_vision_processor(hf_model_path: str):
+    """Monkey-patch KimiK25VisionProcessor.from_dict to avoid duplicate keyword errors.
+
+    The upstream from_dict passes both **config and **kwargs to cls(), which causes
+    'got multiple values for keyword argument' when AutoProcessor injects '_from_auto'.
+    """
+    try:
+        from transformers.dynamic_module_utils import get_class_from_dynamic_module
+
+        klass = get_class_from_dynamic_module(
+            "kimi_k25_vision_processing.KimiK25VisionProcessor",
+            hf_model_path,
+        )
+        if klass is None or getattr(klass, "_from_dict_patched", False):
+            return
+
+        @classmethod  # type: ignore[misc]
+        def _patched_from_dict(cls, config_dict, **kwargs):
+            config = config_dict.copy()
+            for key in list(kwargs.keys()):
+                config.pop(key, None)
+            media_proc_cfg = config.pop("media_proc_cfg", {})
+            return cls(media_proc_cfg=media_proc_cfg, **config, **kwargs)
+
+        klass.from_dict = _patched_from_dict
+        klass._from_dict_patched = True
+    except Exception:
+        pass
+
+
+def pre_expand_vision_tokens(input_ids, grid_thws, vision_token_id, spatial_merge_size=2):
+    """Pre-expand single vision placeholders (image or video) to N placeholders
+    matching vision feature count.
+
+    With PP > 1 the pipeline schedule needs to know the actual sequence length
+    upfront.  Dynamic expansion inside the model changes seq_length during
+    forward, causing send/recv shape mismatches.  Pre-expanding makes the model
+    use the 1:1 replacement path (is_pre_expanded=True).
+    """
+    if grid_thws is None:
+        return input_ids
+
+    feature_counts = []
+    for grid_thw in grid_thws:
+        t, h, w = grid_thw.tolist()
+        num_features = int(t * (h // spatial_merge_size) * (w // spatial_merge_size))
+        feature_counts.append(num_features)
+
+    expanded = []
+    feat_idx = 0
+    for token_id in input_ids[0]:
+        if token_id.item() == vision_token_id and feat_idx < len(feature_counts):
+            expanded.extend([vision_token_id] * feature_counts[feat_idx])
+            feat_idx += 1
+        else:
+            expanded.append(token_id.item())
+
+    return torch.tensor([expanded], dtype=input_ids.dtype, device=input_ids.device)
+
+
+def pad_input_ids_to_tp_multiple(input_ids, tp_size: int, pad_token_id: int = 0):
+    """Pad input_ids so sequence length is divisible by tp_size.
+
+    Needed for sequence-parallel, which is required for MoE models using TP + EP.
+    No-op when tp_size is 1 or the sequence is already aligned.
+    """
+    if tp_size <= 1:
+        return input_ids
+    seq_len = input_ids.shape[1]
+    remainder = seq_len % tp_size
+    if remainder == 0:
+        return input_ids
+    pad_len = tp_size - remainder
+    padding = torch.full((input_ids.shape[0], pad_len), pad_token_id, dtype=input_ids.dtype, device=input_ids.device)
+    return torch.cat([input_ids, padding], dim=1)
+
+
+def load_image(image_path: str) -> Image.Image:
+    """Load an image from URL or file path."""
+    if image_path.startswith(("http://", "https://")):
+        is_safe, reason = is_safe_public_http_url(image_path)
+        if not is_safe:
+            raise ValueError(f"Refusing to fetch image URL ({reason}): {image_path}")
+        with safe_url_open(image_path) as resp:
+            return Image.open(io.BytesIO(resp.read()))
+    return Image.open(image_path)
+
+
+def to_cuda(x):
+    """Move a tensor, list of tensors, or None to CUDA."""
+    if x is None:
+        return None
+    if isinstance(x, (list, tuple)):
+        return [t.cuda() for t in x]
+    return x.cuda()
+
+
+def process_multi_image_inputs(processor, image_paths: list[str], prompt: str):
+    """Process N ordered images + prompt into model inputs (Qwen-family).
+
+    Returns:
+        (input_ids, pixel_values, image_grid_thw)
+    """
+    if not _HAS_QWEN_VL_UTILS:
+        raise ImportError("qwen-vl-utils required: pip install qwen-vl-utils")
+    pils = [load_image(p).convert("RGB") for p in image_paths]
+    messages = [
+        {
+            "role": "user",
+            "content": [{"type": "image", "image": p} for p in pils] + [{"type": "text", "text": prompt}],
+        }
+    ]
+    image_inputs, video_inputs = process_vision_info(messages)
+    text = processor.apply_chat_template(messages, tokenize=False, add_generation_prompt=True)
+    inputs = processor(text=[text], images=image_inputs, videos=video_inputs, padding=True, return_tensors="pt")
+    return inputs.input_ids, inputs.get("pixel_values"), inputs.get("image_grid_thw")
+
+
+def process_video_inputs(processor, video_path: str, prompt: str, *, fps: float = 2.0):
+    """Process a video + prompt into model inputs (Qwen-family).
+
+    Frame decoding mirrors the Qwen3-VL training pipeline: fetch_video decodes at
+    ``fps``, then video_processor is called with do_sample_frames=False to use the
+    pre-decoded frames as-is.
+
+    Returns:
+        (input_ids, pixel_values_videos, video_grid_thw)
+    """
+    if not _HAS_QWEN_VL_UTILS:
+        raise ImportError("qwen-vl-utils required: pip install qwen-vl-utils")
+    from qwen_vl_utils import fetch_video
+
+    frames = fetch_video({"video": video_path, "fps": fps})
+    messages = [
+        {
+            "role": "user",
+            "content": [{"type": "video"}, {"type": "text", "text": prompt}],
+        }
+    ]
+    text = processor.apply_chat_template(messages, tokenize=False, add_generation_prompt=True)
+    text_inputs = processor(text=[text], padding=True, return_tensors="pt")
+    video_proc = processor.video_processor(videos=[frames], return_tensors="pt", do_sample_frames=False)
+    # processor(text=...) without videos produces a single <|video_pad|> placeholder.
+    # Pre-expand to match actual vision feature count so PP send/recv shapes are correct.
+    video_token_id = processor.tokenizer.convert_tokens_to_ids("<|video_pad|>")
+    input_ids = pre_expand_vision_tokens(
+        text_inputs["input_ids"],
+        video_proc["video_grid_thw"],
+        vision_token_id=video_token_id,
+    )
+    return input_ids, video_proc.get("pixel_values_videos"), video_proc.get("video_grid_thw")
+
+
+def process_image_inputs(
+    processor,
+    image_path: Optional[str],
+    prompt: str,
+    *,
+    is_gemma4: bool = False,
+    is_kimi: bool = False,
+    is_mistral3: bool = False,
+    image_token_id: Optional[int] = None,
+):
+    """Process image + prompt into model inputs.
+
+    Returns:
+        (input_ids, pixel_values, image_grid_thw, image_sizes, mm_token_type_ids,
+         image_position_ids)
+
+        Fields not applicable to the current model are None.
+    """
+    if is_gemma4:
+        return _process_gemma4_inputs(processor, image_path, prompt)
+
+    if not image_path:
+        inputs = processor(text=[prompt], return_tensors="pt")
+        return inputs.input_ids, None, None, None, None, None
+
+    if is_kimi:
+        return _process_kimi_inputs(processor, image_path, prompt, image_token_id)
+
+    return _process_default_inputs(processor, image_path, prompt, allow_raw_prompt=is_mistral3)
+
+
+def _process_kimi_inputs(processor, image_path, prompt, image_token_id):
+    messages = [
+        {"role": "system", "content": "You are Kimi, an AI assistant created by Moonshot AI."},
+        {
+            "role": "user",
+            "content": [
+                {"type": "image", "image_url": load_image(image_path)},
+                {"type": "text", "text": prompt},
+            ],
+        },
+    ]
+    inputs = processor(messages=messages)
+    grid_thws = getattr(inputs, "grid_thws", None)
+    input_ids = pre_expand_vision_tokens(inputs.input_ids, grid_thws, image_token_id)
+    return input_ids, inputs.pixel_values, grid_thws, None, None, None
+
+
+def _process_default_inputs(processor, image_path, prompt, *, allow_raw_prompt=False):
+    if getattr(processor, "chat_template", None) is None:
+        if not allow_raw_prompt:
+            raise ValueError("The processor has no chat template and no model-specific raw-prompt fallback")
+        image = load_image(image_path).convert("RGB")
+        image_token = getattr(processor, "image_token", None) or getattr(
+            processor.tokenizer, "image_token", "<|image|>"
+        )
+        if image_token not in prompt:
+            prompt = f"{image_token}\n{prompt}"
+        inputs = processor(
+            text=[prompt],
+            images=[image],
+            padding=True,
+            return_tensors="pt",
+        )
+        return (
+            inputs.input_ids,
+            inputs.get("pixel_values"),
+            inputs.get("image_grid_thw"),
+            inputs.get("image_sizes"),
+            inputs.get("mm_token_type_ids"),
+            inputs.get("image_position_ids"),
+        )
+
+    if not _HAS_QWEN_VL_UTILS:
+        raise ImportError(
+            "qwen_vl_utils is required for chat-template VLM image processing. Install it with: pip install qwen-vl-utils"
+        )
+
+    messages = [
+        {
+            "role": "user",
+            "content": [
+                {"type": "image", "image": image_path},
+                {"type": "text", "text": prompt},
+            ],
+        }
+    ]
+    image_inputs, video_inputs = process_vision_info(messages)
+    text = processor.apply_chat_template(messages, tokenize=False, add_generation_prompt=True)
+    inputs = processor(
+        text=[text],
+        images=image_inputs,
+        videos=video_inputs,
+        padding=True,
+        return_tensors="pt",
+    )
+    return (
+        inputs.input_ids,
+        inputs.get("pixel_values"),
+        inputs.get("image_grid_thw"),
+        inputs.get("image_sizes"),
+        inputs.get("mm_token_type_ids"),
+        inputs.get("image_position_ids"),
+    )
+
+
+def _process_gemma4_inputs(processor, image_path, prompt):
+    """Process inputs for Gemma 4 (text-only or vision+text).
+
+    Uses apply_chat_template for instruction-tuned models; falls back to raw
+    processor with manual image token injection for base models.
+
+    Returns:
+        (input_ids, pixel_values, None, None, None, image_position_ids)
+    """
+    if image_path:
+        image = load_image(image_path)
+        has_chat_template = (
+            hasattr(processor, "apply_chat_template") and getattr(processor, "chat_template", None) is not None
+        )
+        if has_chat_template:
+            messages = [
+                {
+                    "role": "user",
+                    "content": [
+                        {"type": "image", "image": image},
+                        {"type": "text", "text": prompt},
+                    ],
+                }
+            ]
+            inputs = processor.apply_chat_template(
+                messages,
+                tokenize=True,
+                return_dict=True,
+                return_tensors="pt",
+                add_generation_prompt=True,
+            )
+        else:
+            image_token = getattr(processor.tokenizer, "image_token", "<|image|>")
+            if image_token not in prompt:
+                prompt = image_token + "\n" + prompt
+            inputs = processor(text=[prompt], images=[image], return_tensors="pt")
+    else:
+        inputs = processor.tokenizer(text=[prompt], return_tensors="pt")
+
+    return (
+        inputs["input_ids"],
+        inputs.get("pixel_values"),
+        None,
+        None,
+        None,
+        inputs.get("image_position_ids"),
+    )
