+++ tests/unit_tests/scripts/inference/test_setup_inference.py
@@ -124,11 +124,7 @@ def test_parser_forwards_model_checkpoint_prompt_and_engine_args():
     ("task_name", "expected_path"),
     [
         ("text-generation", "/opt/Megatron-Bridge/scripts/inference/text_generation.py"),
-        ("vlm-generation", "/opt/Megatron-Bridge/examples/conversion/hf_to_megatron_generate_vlm.py"),
-        (
-            "model-comparison",
-            "/opt/Megatron-Bridge/examples/conversion/compare_hf_and_megatron/compare.py",
-        ),
+        ("vlm-generation", "/opt/Megatron-Bridge/scripts/inference/vlm_generation.py"),
     ],
 )
 def test_parser_selects_repository_inference_task(task_name, expected_path):
